ut fund inquiry/requirements.txt
google-api-python-client==2.165.0
google-auth==2.39.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
